@Kobzol 2024-10-21
I saw only about ~40 ms improvement locally, but this might be worse for HDDs and/or more complex git index states. Even though this creates a new output function, it does not complicate the code very much and seems like it can have useful perf. benefits. Let's land this for now.

@bors r+