@rust-log-analyzer 2024-10-21
A job failed! Check out the build log: [(web)](https://github.com/rust-lang-ci/rust/actions/runs/11441974113/job/31831295400) [(plain)](https://github.com/rust-lang-ci/rust/commit/39f707ffd1e85c9f426036227bffbc028c784cca/checks/31831295400/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain

```

</details>